requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
macholib==1.16.3
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pytube==15.0.0
setuptools==69.5.1
